refact: separete types to a file
