+++ data/en/services.ts
@@ -1,15 +1,8 @@
-import { IconType } from "react-icons/lib";
-
 import { FaServer } from 'react-icons/fa'
 import { GiSmartphone } from 'react-icons/gi'
 import { MdDesktopWindows } from 'react-icons/md'
 import { MdLaptopMac } from 'react-icons/md'
-
-export interface ServicesInterface {
-  name: string
-  description: string
-  icon?: IconType
-}
+import { ServicesInterface } from '../types'
 
 const services: ServicesInterface[] = [
   {
