+++ data/en/skills.ts
@@ -1,15 +1,8 @@
-import { IconType } from "react-icons/lib";
-
 import { FaReact, FaNodeJs, FaHtml5, FaCss3 } from 'react-icons/fa'
 import { SiNextDotJs, SiGraphql, SiTypescript } from 'react-icons/si'
 import { ImGit } from 'react-icons/im'
 import getExperienceTime from '../../utils/getExperienceTime'
-
-export interface SkillsInterface {
-  name: string
-  expTime: () => string
-  icon?: IconType
-}
+import { SkillsInterface } from '../types'
 
 const skills: SkillsInterface[] = [
   {
